pyproject.toml
[project]
name = "api-tester"
version = "0.1.0"
description = "AI-powered API testing agent using OpenAI Agents SDK"
readme = "README.md"
authors = [
    { name = "waleed260", email = "vkdeku20@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
api-tester = "api_tester:main"

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"
